# _swagger_test_output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Noor API Test",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "paths": {
...
    "/adhkar": {
      "get": {
        "tags": [
          "Adhkar (الأذكار)"
        ],
        "summary": "الشاشة الرئيسية للأذكار (ورد اليوم + 6 فئات)",
        "description": "يُرجع ورد اليوم مع نسبة التقدم (Progress bar) + قائمة كل الفئات الست (الصباح، المساء، النوم، المسجد، الصلاة، ورد اليوم) للعرض في الشاشة الأولى من تبويب الأذكار. المحتوى مصدره حصن المسلم + صحيح البخاري ومسلم. Optional Bearer: when a valid access token is present, dailyWird progress is personalized for that user.\n",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "✅ ورد اليوم + قائمة الفئات",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Dhikr home (categories + daily wird) retrieved successfully",
                  "data": {
                    "greeting": "واذكر ربك إذا نسيت",
                    "dailyWird": {
                      "titleAr": "وردك اليوم",
                      "subtitleAr": "واذكر ربك إذا نسيت",
                      "progressItemsDone": 4,
                      "progressItemsTotal": 8,
                      "progressPercent": 50,
                      "ctaAr": "اكمل وردك اليوم",
                      "categoryKey": "GENERAL_WIRD",
                      "items": [
                        {
                          "id": "uuid",
                          "orderInCategory": 1,
                          "textAr": "لَا حَوْلَ وَلَا قُوَّةَ إِلَّا بِاللَّهِ الْعَلِيِّ الْعَظِيمِ",
                          "repeatCount": 100,
                          "referenceAr": "رواه البخاري ومسلم",
                          "benefitAr": "كنز من كنوز الجنة، ومفتاح لكل باب خير"
                        }
                      ]
                    },
                    "categories": [
                      {
                        "id": "cat-uuid-1",
                        "key": "MORNING",
                        "nameAr": "اذكار الصباح",
                        "nameEn": "Morning Dhikr",
                        "descriptionAr": "الأذكار الواردة لصباح المسلم كل يوم من حصن المسلم",
                        "iconCode": "🌤️",
                        "sortOrder": 1,
                        "totalItems": 12
                      },
                      {
                        "id": "cat-uuid-2",
                        "key": "EVENING",
                        "nameAr": "اذكار المساء",
                        "nameEn": "Evening Dhikr",
                        "iconCode": "🌙",
                        "sortOrder": 2,
                        "totalItems": 11
                      }
                    ]
                  },
                  "meta": null,
                  "timestamp": "2026-08-27T03:15:00.000Z",
                  "requestId": "uuid"
                }
              }
            }
          }
        }
      }
    },
...
    "/adhkar/progress": {
      "get": {
        "tags": [
          "Adhkar (الأذكار)"
        ],
        "summary": "تقدم المستخدم في فئة أذكار معينة لليوم",
        "description": "يُرجع حالة كل ذكر (tapCount + completed) + markedItemId للاستكمال.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoryKey",
            "schema": {
              "type": "string",
              "example": "MORNING"
            },
            "description": "مفتاح الفئة (MORNING, EVENING, BEFORE_SLEEP, ...)"
          }
        ],
        "responses": {
          "200": {
            "description": "✅ تقدم الأذكار",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "data": {
                    "categoryKey": "MORNING",
                    "markedItemId": "item-uuid",
                    "items": [
                      {
                        "itemId": "item-uuid",
                        "tapCount": 2,
                        "completed": false
                      }
                    ],
                    "progressItemsDone": 3,
                    "progressItemsTotal": 20,
                    "progressPercent": 15
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "put": {
        "tags": [
          "Adhkar (الأذكار)"
        ],
        "summary": "حفظ تقدم ذكر معين (tap count)",
        "description": "يحفظ عدد النقرات لذكر معين لليوم الحالي ويُرجع التقدم الكامل.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "categoryKey",
                  "itemId",
                  "tapCount"
                ],
                "properties": {
                  "categoryKey": {
                    "type": "string",
                    "example": "MORNING"
                  },
                  "itemId": {
                    "type": "string",
                    "example": "item-uuid"
                  },
                  "tapCount": {
                    "type": "integer",
                    "example": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "✅ تم حفظ التقدم"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/adhkar/categories/{key}": {
      "get": {
        "tags": [
          "Adhkar (الأذكار)"
        ],
        "summary": "تفاصيل فئة معينة + كل الأذكار فيها (مثلاً اذكار الصباح مفصلين)",
        "description": "يُرجع الفئة + قائمة الأذكار داخلها بالترتيب، مع نص الذكر بالتشكيل الكامل وعدد مرات التكرار (التكرار) + المرجع (رواه البخاري) + فضل الذكر (إن وجد) — كل الأزرار \"مشاركة\" في الشاشة التانية تعتمد على (textAr + repeatCount + referenceAr).\n",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "MORNING",
                "EVENING",
                "BEFORE_SLEEP",
                "ENTERING_MOSQUE",
                "AFTER_PRAYER",
                "GENERAL_WIRD",
                "TRAVEL",
                "SICK",
                "FOOD",
                "ISTIKHARA",
                "WUDU",
                "ISTIGHFAR",
                "QAYN",
                "MASJID_AFTER_SALAM"
              ],
              "example": "MORNING"
            },
            "description": "مفتاح الفئة (حرف كبير أو صغير - الـ backend يحول لحروف كبير)"
          }
        ],
        "responses": {
          "200": {
            "description": "✅ تفاصيل الفئة مع الأذكار",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Dhikr category MORNING retrieved successfully",
                  "data": {
                    "id": "cat-uuid-morning",
                    "key": "MORNING",
                    "nameAr": "اذكار الصباح",
                    "nameEn": "Morning Dhikr",
                    "descriptionAr": "الأذكار الواردة لصباح المسلم كل يوم من حصن المسلم",
                    "iconCode": "🌤️",
                    "sortOrder": 1,
                    "totalItems": 3,
                    "items": [
                      {
                        "id": "item-uuid-1",
                        "orderInCategory": 1,
                        "textAr": "أَعُوذُ بِاللَّهِ مِنَ الشَّيْطَانِ الرَّجِيمِ. اللَّهُ لَا إِلَٰهَ إِلَّا هُوَ الْحَيُّ الْقَيُّومُ ...\n",
                        "repeatCount": 1,
                        "referenceAr": "آية الكرسي - سورة البقرة 255",
                        "benefitAr": "من قالها حين يصبح أجير من الجن حتى يمسي"
                      },
                      {
                        "id": "item-uuid-2",
                        "orderInCategory": 2,
                        "textAr": "قُلْ هُوَ ٱللَّهُ أَحَدٌ ... (المعوذات)",
                        "repeatCount": 3,
                        "referenceAr": "المعوذات ثلاث",
                        "benefitAr": "من قرأهن ثلاثاً كفتاه من كل شيء"
                      }
                    ]
                  },
                  "meta": null,
                  "timestamp": "2026-08-27T03:15:00.000Z",
                  "requestId": "uuid"
                }
              }
            }
          },
          "404": {
            "description": "❌ الفئة غير موجودة",
            "content": {
              "application/json": {
                "example": {
                  "success": false,
                  "message": "Dhikr category not found for key: XYZ",
                  "code": "NOT_FOUND",
                  "timestamp": "2026-08-27T03:15:00.000Z"
                }
              }
            }
          }
        }
      }
    },
    "/adhkar/favorites": {
      "get": {
        "tags": [
          "Adhkar (الأذكار)"
        ],
        "summary": "قائمة الأذكار المفضلة للمستخدم",
        "description": "يُرجع كل الأذكار التي حفظها المستخدم في المفضلة مع تفاصيل كل ذكر.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "✅ قائمة المفضلة",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Adhkar favorites retrieved successfully",
                  "data": [
                    {
                      "id": "fav-uuid",
                      "itemId": "item-uuid",
                      "dhikr": {
                        "id": "item-uuid",
                        "textAr": "لَا حَوْلَ وَلَا قُوَّةَ إِلَّا بِاللَّهِ",
                        "repeatCount": 100,
                        "referenceAr": "رواه البخاري",
                        "benefitAr": "كنز من كنوز الجنة",
                        "category": {
                          "key": "MORNING",
                          "nameAr": "اذكار الصباح"
                        }
                      },
                      "createdAt": "2026-08-28T10:00:00.000Z"
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Adhkar (الأذكار)"
        ],
        "summary": "إضافة ذكر للمفضلة",
        "description": "حفظ ذكر معين في قائمة المفضلة للمستخدم.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "itemId"
                ],
                "properties": {
                  "itemId": {
                    "type": "string",
                    "example": "item-uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "✅ تمت الإضافة للمفضلة",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Dhikr added to favorites",
                  "data": {
                    "id": "fav-uuid",
                    "itemId": "item-uuid",
                    "dhikr": {
                      "id": "item-uuid",
                      "textAr": "لَا حَوْلَ وَلَا قُوَّةَ إِلَّا بِاللَّهِ",
                      "repeatCount": 100,
                      "referenceAr": "رواه البخاري",
                      "category": {
                        "key": "MORNING",
                        "nameAr": "اذكار الصباح"
                      }
                    },
                    "createdAt": "2026-08-28T10:00:00.000Z"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "409": {
            "description": "❌ الذكر موجود في المفضلة"
          }
        }
      }
    },
    "/adhkar/favorites/{favoriteId}": {
      "delete": {
        "tags": [
          "Adhkar (الأذكار)"
        ],
        "summary": "إزالة ذكر من المفضلة",
        "description": "حذف ذكر من قائمة المفضلة.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "favoriteId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "معرف المفضلة (favorite ID)"
          }
        ],
        "responses": {
          "200": {
            "description": "✅ تمت الإزالة",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "Favorite removed successfully",
                  "data": {
                    "message": "Favorite removed successfully"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "description": "❌ المفضلة غير موجودة"
          }
        }
      }
    },
...
    "/journey/adhkar": {
      "patch": {
        "tags": [
          "Journey"
        ],
        "summary": "تبديل حالة إكمال أذكار اليوم",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "completed"
                ],
                "properties": {
                  "completed": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "examples": {
                "default": {
                  "summary": "وضع علامة أن أذكار اليوم مكتملة",
                  "value": {
                    "completed": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "✅ تم التحديث",
            "content": {
              "application/json": {
                "example": {
                  "success": true,
                  "message": "تم تحديث حالة الأذكار بنجاح",
                  "data": {
                    "date": "2026-07-27",
                    "morningCompleted": true,
                    "eveningCompleted": true,
                    "overallCompleted": true,
                    "percent": 100
                  },
                  "meta": null,
                  "timestamp": "2026-07-27T10:30:00.000Z"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
...
}